XMonad development setup
I was too annoyed by the fact that the Xmonad config that I had did not
have any IDE support and direnv + nix-direnv + flake/nix-shell offers a
really great way to set this up. Now I have editor support. YAY!

One thing that bothers me still is the fact that all these are not the
same packages that the system uses (or they might be I am not sure). Not
really sure what is the way to solve this problem though.